+++ src/constants/ItemConstants.java
@@ -35,16 +35,16 @@ import java.util.Map;
 public final class ItemConstants {
     protected static Map<Integer, MapleInventoryType> inventoryTypeCache = new HashMap<>();
     
-    public final static int LOCK = 0x01;
-    public final static int SPIKES = 0x02;
-    public final static int KARMA_USE = 0x02;
-    public final static int COLD = 0x04;
-    public final static int UNTRADEABLE = 0x08;
-    public final static int KARMA_EQP = 0x10;
-    public final static int SANDBOX = 0x40;             // let 0x40 until it's proven something uses this
-    public final static int PET_COME = 0x80;
-    public final static int ACCOUNT_SHARING = 0x100;
-    public final static int MERGE_UNTRADEABLE = 0x200;
+    public final static short LOCK = 0x01;
+    public final static short SPIKES = 0x02;
+    public final static short KARMA_USE = 0x02;
+    public final static short COLD = 0x04;
+    public final static short UNTRADEABLE = 0x08;
+    public final static short KARMA_EQP = 0x10;
+    public final static short SANDBOX = 0x40;             // let 0x40 until it's proven something uses this
+    public final static short PET_COME = 0x80;
+    public final static short ACCOUNT_SHARING = 0x100;
+    public final static short MERGE_UNTRADEABLE = 0x200;
 
     public final static boolean EXPIRING_ITEMS = true;
     public final static Set<Integer> permanentItemids = new HashSet<>();
@@ -147,7 +147,7 @@ public final class ItemConstants {
         return scrollId == 2040727 || scrollId == 2041058;
     }
     
-    public static boolean isFlagModifier(int scrollId, byte flag) {
+    public static boolean isFlagModifier(int scrollId, short flag) {
         if(scrollId == 2041058 && ((flag & ItemConstants.COLD) == ItemConstants.COLD)) return true;
         if(scrollId == 2040727 && ((flag & ItemConstants.SPIKES) == ItemConstants.SPIKES)) return true;
         return false;
